Check for the availability of a scripting engine.
A scripting engine might not be available, because certain internal operations in the early phase of starting an application (e.g. loading sessions) might not have an associated request evaluator yet.

I have decided against having a fully initialized request evaluator ready and associated when loading persistent sessions, as this always was like this and the need to have a scripting engine and thus a request evaluator when committing transactions was only introduced lately by adding several commit callbacks. Not having a scripting engine at hand now only means the callbacks won't be called, which always was like this anyways and seems acceptable in the early phase of apllication startup.